+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is inspired by [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [0.5.1] - 2025-12-31
+
+### Changed
+
+- Standardized email authentication result structs to match wire format and other SDKs
+
+### Added
+
+- End-to-end integration tests for email authentication results using the test email API
+
 ## [0.5.0] - 2025-12-08
 
 ### Initial release